platfr/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
environs==11.2.1
gunicorn==23.0.0
django-crispy-forms==2.0